[T159] bug fix open gallery not working (#112)
the issue was that on my test device which is a pixel 2, open gallery
was working but in higher android versions, open gallery was not working
and this is why:

### _In Android, the need to request permission to access the device's
gallery depends on the Android version and the specific storage access
framework used by the app._

**Scoped Storage (Android 10+):** On Android 10 and newer versions,
Google introduced scoped storage. Under scoped storage, apps are limited
in how they can access files on the device, and they are encouraged to
use system-provided file pickers to access files from the user's device.
In this case, you typically don't need to request permission to access
the gallery because the user selects the files via the system file
picker, which is granted permission to access the files.

**Legacy Storage (Android 9 and older):** On Android versions prior to
Android 10, you typically need to request permission to access the
user's gallery as apps had broad access to the file system. This
required the READ_EXTERNAL_STORAGE permission to access external
storage.